+++ client/src/lib.rs
@@ -8,7 +8,7 @@
 // TODO: Handshake
 // TODO: O(1) comparison for IDS.
 // TODO: Remove #[allow(dead_code)]s.
-// TODO: Add ability to name snapshots
+// TODO: ~~Add ability to name snapshots~~
 // TODO: Add ability to read and revoke records not on the head of the chain.
 // TODO: Add Reference types for the RecordIds and VaultIds to expose to the External programs.